Refactored Welcome component
