+++ src/components/Welcome.client.tsx
@@ -1,0 +1,12 @@
+export default function Welcome() {
+  return (
+    <div className="text-gray-900 pt-16 rounded-[40px] my-8 px-4 xl:px-12 -mx-4 xl:-mx-12">
+      <div className="text-center mb-16">
+        <h1 className="font-sansSerif font-semibold mb-4 text-shadow-xl text-white text-5xl leading-tight">
+          This valentine&#39;s Day, get your sweetheart just what they&#39;ve
+          wanted (at just the right price).
+        </h1>
+      </div>
+    </div>
+  );
+}
